% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Execute complete enrichment analysis (hypergeometric test, BH correction).}
\usage{
enrichment_analysis(gene_list, tissue_df, alpha = 0.05, filename = NULL)
}
\arguments{
\item{gene_list}{a list of non-redundant Wormbase IDs.}

\item{alpha}{significance threshold, defaults to 0.05.}

\item{filename}{if not NULL, file name to save results as a csv.}

\item{dictionary}{dictionary created with `fetch_dictionary()`.}
}
\value{
a data.frame containing significantly enriched tissues
}
\description{
Execute complete enrichment analysis (hypergeometric test, BH correction).
}
