% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tea.R
\name{install_tea}
\alias{install_tea}
\title{Install the tissue enrichment analysis Python package}
\usage{
install_tea(..., envname = "r-tea")
}
\arguments{
\item{...}{Additional arguments to be passed to `reticulate::py_install()`}

\item{envname}{Name for a Python virtual environment in which to perform install}
}
\description{
Install the tissue enrichment analysis Python package
}
