% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Execute complete enrichment analysis (hypergeometric test, BH correction).}
\usage{
enrichment_analysis(
  gene_list,
  dictionary,
  alpha = 0.05,
  background_genes = NULL,
  filter = TRUE
)
}
\arguments{
\item{gene_list}{a list of Wormbase IDs.}

\item{dictionary}{dictionary created with `fetch_dictionary()`.}

\item{alpha}{significance threshold.}

\item{background_genes}{list of background genes}

\item{filter}{if TRUE, filters results before returning (mimicking the Python package). Set to FALSE to perform additional FDR corrections.}
}
\value{
a data.frame containing significantly enriched tissues
}
\description{
Execute complete enrichment analysis (hypergeometric test, BH correction).
}
