% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dictionary.R
\name{fetch_dictionary}
\alias{fetch_dictionary}
\title{Fetch a dictionary}
\usage{
fetch_dictionary(analysis = c("tissue", "phenotype", "go"))
}
\arguments{
\item{analysis}{the type of analysis (tissue, phenotype, go)}
}
\value{
A data.frame containing the dictionary of interest, one row per gene, one column per term
}
\description{
Download a dictionary file from Wormbase
}
